test: add missing test

+++ src/main/java/com/octo/masterclass/api/persistence/Plat.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Plat {
         @Id
         @GeneratedValue(strategy= GenerationType.AUTO)
-        int id;
+        Long id;
 
         @Column
         String nom;
@@ -27,11 +27,11 @@ public class Plat {
         public Plat() {
         }
 
-        public int getId() {
+        public Long getId() {
                 return id;
         }
 
-        public void setId(int id) {
+        public void setId(Long id) {
                 this.id = id;
         }
